add fix to linked_list, add bst introduction, removed alerts from linked list in favour of error window, add new errors in errors class, add generate random list to linked list, add new introduction images and text for linked list, fix to main menu resize

+++ AlgoLearn/src/algolearn/gui/algorithms/rbt/BST.java
@@ -26,7 +26,7 @@ public class BST<E extends Comparable<E>> implements Tree<E> {
             return false;
         else if(e.compareTo(root.element) == 0)
         {
-        	System.out.println((int)e+" found!");
+        	System.out.println(e+" found!");
         	path.getElements().add(new LineTo(root.x,root.y));
         	return true;
         }
